layoutの統一
