+++ app/layout.tsx
@@ -1,3 +1,5 @@
+import Footer from '@/components/common/Footer'
+import Header from '@/components/common/Header'
 import { GeistSans } from 'geist/font/sans'
 import './globals.css'
 import { Providers } from './provider'
@@ -8,8 +10,8 @@ const defaultUrl = process.env.VERCEL_URL
 
 export const metadata = {
   metadataBase: new URL(defaultUrl),
-  title: 'Next.js and Supabase Starter Kit',
-  description: 'The fastest way to build apps with Next.js and Supabase',
+  title: 'Ingageクイズ',
+  description: '白熱したクイズバトルをあなたにお届けします。',
 }
 
 export default function RootLayout({
@@ -20,7 +22,13 @@ export default function RootLayout({
   return (
     <html lang='en' className={GeistSans.className}>
       <body className='bg-background text-foreground'>
-        <Providers>{children}</Providers>
+        <Providers>
+          <div className='flex flex-col min-h-[100dvh]'>
+            <Header />
+            {children}
+            <Footer />
+          </div>
+        </Providers>
       </body>
     </html>
   )
