+++ app/master/[roomId]/final-results/page.tsx
@@ -1,19 +1,6 @@
-/**
- * v0 by Vercel.
- * @see https://v0.dev/t/V0N8EST8NUN
- * Documentation: https://v0.dev/docs#integrating-generated-code-into-your-nextjs-app
- */
 export default function Component() {
   return (
     <div className='min-h-screen bg-background'>
-      <header className='bg-[#d42f2f] py-4'>
-        <div className='container mx-auto flex items-center justify-center'>
-          <TrophyIcon className='h-8 w-8 text-background' />
-          <h1 className='ml-4 text-2xl font-bold text-background'>
-            Final Results
-          </h1>
-        </div>
-      </header>
       <main className='container mx-auto py-8'>
         <section className='mb-8'>
           <div className='flex items-center justify-center bg-[#d42f2f] p-8 rounded-lg'>
@@ -92,55 +79,5 @@ export default function Component() {
         </section>
       </main>
     </div>
-  )
-}
-
-interface IconProps extends React.SVGProps<SVGSVGElement> {}
-
-function SnowflakeIcon(props: IconProps) {
-  return (
-    <svg
-      {...props}
-      xmlns='http://www.w3.org/2000/svg'
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      stroke='currentColor'
-      strokeWidth='2'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-    >
-      <line x1='2' x2='22' y1='12' y2='12' />
-      <line x1='12' x2='12' y1='2' y2='22' />
-      <path d='m20 16-4-4 4-4' />
-      <path d='m4 8 4 4-4 4' />
-      <path d='m16 4-4 4-4-4' />
-      <path d='m8 20 4-4 4 4' />
-    </svg>
-  )
-}
-
-function TrophyIcon(props: IconProps) {
-  return (
-    <svg
-      {...props}
-      xmlns='http://www.w3.org/2000/svg'
-      width='24'
-      height='24'
-      viewBox='0 0 24 24'
-      fill='none'
-      stroke='currentColor'
-      strokeWidth='2'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-    >
-      <path d='M6 9H4.5a2.5 2.5 0 0 1 0-5H6' />
-      <path d='M18 9h1.5a2.5 2.5 0 0 0 0-5H18' />
-      <path d='M4 22h16' />
-      <path d='M10 14.66V17c0 .55-.47.98-.97 1.21C7.85 18.75 7 20.24 7 22' />
-      <path d='M14 14.66V17c0 .55.47.98.97 1.21C16.15 18.75 17 20.24 17 22' />
-      <path d='M18 2H6v7a6 6 0 0 0 12 0V2Z' />
-    </svg>
   )
 }
