+++ app/master/[roomId]/question/page.tsx
@@ -8,18 +8,6 @@
 export default function Component() {
   return (
     <div className='flex flex-col min-h-screen bg-background'>
-      <header className='flex items-center justify-between px-8 py-4 border-b'>
-        <div className='flex items-center gap-2'>
-          <PuzzleIcon className='w-8 h-8 text-primary' />
-          <h1 className='text-2xl font-bold tracking-tight text-foreground'>
-            Trivia Blitz
-          </h1>
-        </div>
-        <div className='flex items-center gap-2 text-muted-foreground'>
-          <ClockIcon className='w-4 h-4' />
-          <span>00:45</span>
-        </div>
-      </header>
       <main className='flex-1 flex flex-col items-center justify-center px-8 py-16'>
         <div className='max-w-3xl w-full bg-card rounded-lg shadow-lg overflow-hidden'>
           <div className="bg-[url('/geometric-pattern.svg')] bg-cover bg-center bg-no-repeat p-8">
@@ -29,22 +17,6 @@ export default function Component() {
           </div>
         </div>
       </main>
-      <footer className='flex items-center justify-between px-8 py-4 border-t'>
-        <div className='px-4 py-2 rounded bg-gray-200 text-gray-700'>
-          End Quiz
-        </div>
-        <div className='flex items-center gap-2'>
-          <div className='px-4 py-2 rounded bg-gray-200 text-gray-700'>
-            Previous
-          </div>
-          <div className='px-4 py-2 rounded bg-blue-500 text-white'>
-            Next Question
-          </div>
-          <div className='px-4 py-2 rounded bg-gray-200 text-gray-700'>
-            View Results
-          </div>
-        </div>
-      </footer>
     </div>
   )
 }
