+++ app/master/rooms/[id]/page.tsx
@@ -1,38 +1,8 @@
-/**
- * This code was generated by v0 by Vercel.
- * @see https://v0.dev/t/zWyBxpEeMDv
- * Documentation: https://v0.dev/docs#integrating-generated-code-into-your-nextjs-app
- */
-
-/** Add fonts into your Next.js project:
-
-import { Inter } from 'next/font/google'
-
-inter({
-  subsets: ['latin'],
-  display: 'swap',
-})
-
-To read more about using these font, please visit the Next.js documentation:
-- App Directory: https://nextjs.org/docs/app/building-your-application/optimizing/fonts
-- Pages Directory: https://nextjs.org/docs/pages/building-your-application/optimizing/fonts
-**/
-import Link from 'next/link'
 import type { SVGProps } from 'react'
 
 export default function EditRoom() {
   return (
     <div className='min-h-screen bg-white'>
-      <header className='flex items-center justify-between px-4 py-2 bg-[#d42f2f] text-white'>
-        <div className='flex items-center'>
-          <PuzzleIcon className='w-8 h-8 mr-2' />
-          <span className='text-lg font-bold'>QuizApp</span>
-        </div>
-        <Link href='#' className='text-white hover:underline' prefetch={false}>
-          <ArrowLeftIcon className='w-6 h-6' />
-          <span className='sr-only'>Back to Home</span>
-        </Link>
-      </header>
       <main className='max-w-3xl mx-auto py-8 px-4'>
         <h1 className='text-3xl font-bold text-center mb-8 text-[#d42f2f]'>
           Edit Quiz Room
