+++ app/protected/page.tsx
@@ -1,8 +1,8 @@
-import DeployButton from '@/components/DeployButton'
 import AuthButton from '@/components/AuthButton'
-import { createClient } from '@/utils/supabase/server'
-import FetchDataSteps from '@/components/tutorial/FetchDataSteps'
+import DeployButton from '@/components/DeployButton'
 import Header from '@/components/Header'
+import FetchDataSteps from '@/components/tutorial/FetchDataSteps'
+import { createClient } from '@/utils/supabase/server'
 import { redirect } from 'next/navigation'
 
 export default async function ProtectedPage() {
@@ -38,20 +38,6 @@ export default async function ProtectedPage() {
           <FetchDataSteps />
         </main>
       </div>
-
-      <footer className='w-full border-t border-t-foreground/10 p-8 flex justify-center text-center text-xs'>
-        <p>
-          Powered by{' '}
-          <a
-            href='https://supabase.com/?utm_source=create-next-app&utm_medium=template&utm_term=nextjs'
-            target='_blank'
-            className='font-bold hover:underline'
-            rel='noreferrer'
-          >
-            Supabase
-          </a>
-        </p>
-      </footer>
     </div>
   )
 }
